% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causaljt.R
\name{causaljt}
\alias{causaljt}
\title{Causal estimators}
\usage{
causaljt(
  Y,
  D,
  X,
  design = c("rct", "observational"),
  method = c("meandiff", "ols", "directmethod", "ipw", "lr"),
  estimand = c("ate", "att"),
  MLreg = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  CF = TRUE,
  K = 5,
  csplot = FALSE,
  cvps = FALSE,
  cvreg = FALSE,
  Kcv = 5,
  trimm = NULL
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{X}{Tibble with circumstances.}

\item{design}{Whether data comes from a Randomized control trial or
from observational data}

\item{method}{IPW or AIPW}

\item{estimand}{Population over which we want the treatment effect (only
relevant for observational studies)}

\item{MLreg}{machine learner choice for outcome regressions}

\item{MLps}{machine learner choice for propensity scores}

\item{CF}{whether to use Cross-Fitting}

\item{K}{Number of folds in cross-fitting}

\item{csplot}{Plot histograms of propensity score by treatment status}

\item{cvps}{Whether machine learning employed for propensity score should
be chosen by cross-validation}

\item{cvreg}{Whether machine learning employed for outcome regression should
be chosen by cross-validation}

\item{Kcv}{number of folds for cross-validation}

\item{trimm}{two-dimensional vector with lower and upper bound for trimming
low and high propensity scores}
}
\value{
A list with the output and a figure.
}
\description{
Causal estimators
}
