% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqewm4.R
\name{ineqewm4}
\alias{ineqewm4}
\title{Social Welfare Function maximization}
\usage{
ineqewm4(
  Y,
  D,
  X,
  targetX,
  rule = c("tree", "linscore"),
  approx = FALSE,
  approxmatrix = NULL,
  WF = c("ineq", "IOp", "IGM", "util"),
  tigm = 0,
  X1 = NULL,
  quants,
  design = c("rct", "observational"),
  est_method = c("PI", "LR"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  MLalpha = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  CF = FALSE,
  K = 5,
  MLiop = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  parallel = FALSE
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{X}{Tibble with circumstances.}

\item{targetX}{Variables to use for treatment allocation}

\item{rule}{Family of treatment rules to search from}

\item{WF}{Welfare function}

\item{tigm}{Target for Kendall-tau}

\item{X1}{variable to compute Kendall-tau with}

\item{quants}{how many quantiles to split continuous vars in targetX}

\item{ML}{Choice of Machine Learner (for IOp)}
}
\value{
A list with the output and a figure.
}
\description{
Social Welfare Function maximization
}
