% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqewm7.R
\name{ineqewm7}
\alias{ineqewm7}
\title{Social Welfare Function maximization}
\usage{
ineqewm7(
  scores11,
  scores10,
  scores01,
  scores00,
  welfare = c("ineq", "iop", "igm", "util"),
  t = 0,
  targetX,
  rule = c("tree", "linear_rule"),
  quants = NULL,
  verbose_tree = 1
)
}
\arguments{
\item{scores11}{Scores when (i,j) both not treated}

\item{scores10}{Scores when i treated and j not treated}

\item{scores01}{Scores when i not treated and j treated}

\item{welfare}{Welfare function}

\item{t}{Target for Kendall-tau}

\item{targetX}{Variables to use for treatment allocation}

\item{rule}{Family of treatment rules to search from}

\item{quants}{how many quantiles to split continuous vars in targetX}

\item{verbose_tree}{level of progress messaging. 0 none, 1 information
about the root node, 2 information about root node and first subnode,
3 information about root node and both subnodes}
}
\value{
A list with the output and a figure.
}
\description{
Social Welfare Function maximization
}
