% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqewm8.R
\name{ineqewm8}
\alias{ineqewm8}
\title{Social Welfare Function maximization}
\usage{
ineqewm8(
  scores11 = NULL,
  scores10 = NULL,
  scores01 = NULL,
  scores00 = NULL,
  D = NULL,
  ate_scores = NULL,
  welfare = c("ineq", "iop", "igm", "util"),
  t = 0,
  targetX = NULL,
  rule = c("tree", "linear_rule"),
  depth = 2,
  quants = NULL,
  verbose = 1,
  targetXnames_pretty = NULL,
  roundvalues = FALSE,
  roundcate = 2
)
}
\arguments{
\item{scores11}{Scores when (i,j) both not treated}

\item{scores10}{Scores when i treated and j not treated}

\item{scores01}{Scores when i not treated and j treated}

\item{D}{Treatment indicator, used to compute the fraction of
inidividuals actually treated in the data within the groups
generated by the treatment rule}

\item{ate_scores}{the AIPW scores used to estimate the ATE. Used to compute
the CATE in the groups created by the treatment rules.}

\item{welfare}{Welfare function}

\item{t}{Target for Kendall-tau}

\item{targetX}{Variables to use for treatment allocation}

\item{rule}{Family of treatment rules to search from}

\item{quants}{vector indicating how many quantiles to split each var
in targetX. If 0 the var is not split, i.e. quants = c(0,4) would
leave the first var (column) in targetX unchanged and split the second
one into quartiles.}

\item{verbose}{level of progress messaging. For trees: 0 none, 1 information
about the root node, 2 information about root node and first subnode,
3 information about root node and both subnodes. For linear rules: 0 none,
1 information about first coordinate, 2 information about both coordinates.}

\item{targetXnames_pretty}{string vector with the names we wish to see in the
output tree (e.g Mother education instead of educm).}

\item{roundvalues}{whether we want the values of the cutting points in the tree to
appear rounded in the tree}

\item{roundcate}{To how many decimals do we wish to round the CATE of the groups
generated by the treatment rule.}
}
\value{
A list with the output and a figure.
}
\description{
Social Welfare Function maximization
}
