% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqewm_rule.R
\name{ineqewm_rule}
\alias{ineqewm_rule}
\title{Social Welfare Function maximization}
\usage{
ineqewm_rule(
  scores11 = NULL,
  scores10 = NULL,
  scores01 = NULL,
  scores00 = NULL,
  rule = NULL,
  welfare = c("ineq", "iop", "igm", "util"),
  t = 0
)
}
\arguments{
\item{scores11}{Scores when (i,j) both not treated}

\item{scores10}{Scores when i treated and j not treated}

\item{scores01}{Scores when i not treated and j treated}

\item{rule}{rule for which we want to compute welfare.}

\item{welfare}{Welfare function}

\item{t}{Target for Kendall-tau}
}
\value{
Estimated welfare and Kendall-tau for IGM case.
}
\description{
Social Welfare Function maximization
}
