% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqewmcpp.R
\name{ineqewmcpp}
\alias{ineqewmcpp}
\title{Social Welfare Function maximization}
\usage{
ineqewmcpp(
  scores11,
  scores10,
  scores01,
  scores00,
  targetX,
  rule = c("tree", "linscore"),
  approx = FALSE,
  approxmatrix = NULL
)
}
\arguments{
\item{targetX}{Variables to use for treatment allocation}

\item{rule}{Family of treatment rules to search from}

\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{X}{Tibble with circumstances.}

\item{WF}{Welfare function}

\item{tigm}{Target for Kendall-tau}

\item{X1}{variable to compute Kendall-tau with}

\item{quants}{how many quantiles to split continuous vars in targetX}

\item{ML}{Choice of Machine Learner (for IOp)}
}
\value{
A list with the output and a figure.
}
\description{
Social Welfare Function maximization
}
