% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wigm.R
\name{wigm}
\alias{wigm}
\title{Kendall tau Welfare Function maximization}
\usage{
wigm(Y, X1, D, t, rule)
}
\arguments{
\item{Y}{Outcome vector.}

\item{X1}{variable to compute Kendall-tau with}

\item{D}{Treatment assignment.}

\item{t}{target Kendall-tau}

\item{rule}{Treatment rule.}
}
\value{
A list with the output and a figure.
}
\description{
Kendall tau Welfare Function maximization
}
