% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wigm_scores.R
\name{wigm_scores}
\alias{wigm_scores}
\title{Kendall-tau Welfare Function maximization}
\usage{
wigm_scores(
  Y,
  X1,
  D,
  X,
  pscore,
  t,
  design = c("rct", "observational"),
  est_method = c("PI", "LR"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  MLalpha = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  CF = TRUE,
  K = 5
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{X1}{outcome with which to compute Kendall-tau with}

\item{D}{Treatment assignment.}

\item{t}{Target for Kendall-tau}

\item{design}{Whether data comes from an RCT or from observational data}

\item{est_method}{Whether traditional plug in estimators are to be used
or locally robust estimators}

\item{MLps}{Choice of machine learner for propensity score}

\item{MLalpha}{Choice of machine learner for additional nuisance parameter}

\item{CF}{Whether to use Cross-fitting}

\item{K}{Number of folds in Cross-fitting}
}
\value{
A list with the output and a figure.
}
\description{
Kendall-tau Welfare Function maximization
}
