% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineq.R
\name{wineq}
\alias{wineq}
\title{Standard Gini Social Welfare Function maximization}
\usage{
wineq(Y, D, rule)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{rule}{Treatment rule.}
}
\value{
A list with the output and a figure.
}
\description{
Standard Gini Social Welfare Function maximization
}
