% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineq_scores.R
\name{wineq_scores}
\alias{wineq_scores}
\title{Standard Gini Social Welfare score computation}
\usage{
wineq_scores(
  Y,
  D,
  X,
  design = c("rct", "observational"),
  est_method = c("PI", "LR"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  MLalpha = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  CF = TRUE,
  K = 5
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{design}{Whether data comes from an RCT or from observational data}

\item{est_method}{Whether traditional plug in estimators are to be used
or locally robust estimators}

\item{MLps}{Choice of machine learner for propensity score}

\item{MLalpha}{Choice of machine learner for additional nuisance parameter}

\item{CF}{Whether to use Cross-fitting}

\item{K}{Number of folds in Cross-fitting}
}
\value{
A matrix with scores and id of the pair
}
\description{
Standard Gini Social Welfare score computation
}
