% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wineq_scores_dyad.R
\name{wineq_scores_dyad}
\alias{wineq_scores_dyad}
\title{Standard Gini Social Welfare score computation}
\usage{
wineq_scores_dyad(
  Y,
  D,
  X,
  pscore,
  design = c("rct", "observational"),
  est_method = c("PI", "LR"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  MLineq = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  CF = TRUE,
  K = 5,
  leaveoneout = FALSE,
  verbose = 1
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{X}{Covariates.}

\item{pscore}{propensity score (use if known, e.g. in an RCT)}

\item{design}{Whether data comes from an RCT or from observational data}

\item{est_method}{Whether traditional plug in estimators are to be used
or locally robust estimators}

\item{MLps}{Choice of machine learner for propensity score}

\item{MLineq}{Choice of machine learner for dyadic regression}

\item{CF}{Whether to use Cross-fitting}

\item{K}{Number of folds in Cross-fitting}
}
\value{
A matrix with scores and id of the pair
}
\description{
Standard Gini Social Welfare score computation
}
