% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiop.R
\name{wiop}
\alias{wiop}
\title{Inequality of Opportunity Social Welfare Function maximization}
\usage{
wiop(
  Y,
  D,
  X,
  rule,
  est_method = c("PI", "LR"),
  MLiop = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  CF = TRUE,
  K = 5
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{X}{Tibble with circumstances.}

\item{rule}{Treatment rule.}

\item{ML}{Choice of Machine Learner}
}
\value{
A list with the output and a figure.
}
\description{
Inequality of Opportunity Social Welfare Function maximization
}
