% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiop_scores.R
\name{wiop_scores}
\alias{wiop_scores}
\title{Inequality of Opportunity Social Welfare Function maximization}
\usage{
wiop_scores(
  Y,
  D,
  X,
  design = c("rct", "observational"),
  est_method = c("PI", "LR"),
  MLiop = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  MLps = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  MLalpha = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "SL"),
  CF = TRUE,
  K = 5
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{X}{Tibble with circumstances.}

\item{design}{Whether data comes from an RCT or from observational data}

\item{est_method}{Whether traditional plug in estimators are to be used
or locally robust estimators}

\item{MLiop}{Choice of Machine Learner for outcome regression}

\item{MLps}{Choice of Machine Learner for propensity score}
}
\value{
A list with the output and a figure.
}
\description{
Inequality of Opportunity Social Welfare Function maximization
}
