% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsims.R
\name{wsims}
\alias{wsims}
\title{True ewm simulation}
\usage{
wsims(Y1, Y0, fv1, fv0, rule, X1, t, WF = c("ineq", "IOp", "IGM", "util"))
}
\arguments{
\item{Y1}{Potential outcome under treatment}

\item{Y0}{Potential outcome under no treatment}

\item{fv1}{Potential fitted values under treatment}

\item{fv0}{Potential fitted values under no treatment}

\item{rule}{Treatment rule.}

\item{X1}{outcome with which to compute Kendall-tau with}

\item{t}{Target for Kendall-tau}

\item{WF}{Welfare function}
}
\value{
A list with the output and a figure.
}
\description{
True ewm simulation
}
