% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wutil.R
\name{wutil}
\alias{wutil}
\title{Utilitarian Social Welfare Function maximization}
\usage{
wutil(Y, D, rule)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment assignment.}

\item{rule}{Treatment rule.}
}
\value{
A list with the output and a figure.
}
\description{
Utilitarian Social Welfare Function maximization
}
